fixing warning messages in latestw_all

+++ contrib/win32/win32compat/fileio.c
@@ -781,7 +781,7 @@ fileio_stat(const char *path, struct _stat64 *buf)
 		goto cleanup;
 	}
 	
-	len = wcslen(wpath);
+	len = (int)wcslen(wpath);
 
 	buf->st_ino = 0; /* Has no meaning in the FAT, HPFS, or NTFS file systems*/
 	buf->st_gid = 0; /* UNIX - specific; has no meaning on windows */
